zap actions.py -> move classes to subject.py
